Rename Components, Add Staff to Section Entity
