+++ frontend/src/app/academics/academics-home/academics-home.component.html
@@ -1,0 +1,1 @@
+<p>academics-home works!</p>
