+++ frontend/src/app/academics/academics-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CoursesHomeComponent } from './courses-home/courses-home.component';
+import { CoursesHomeComponent } from './course-catalog/course-catalog.component';
 import { OfferingsComponent } from './offerings/offerings.component';
 
 const routes: Routes = [CoursesHomeComponent.Route, OfferingsComponent.Route];
