+++ frontend/src/app/academics/academics.module.ts
@@ -8,15 +8,20 @@ import { MatExpansionModule } from '@angular/material/expansion';
 import { MatButtonModule } from '@angular/material/button';
 import { MatTableModule } from '@angular/material/table';
 import { AcademicsRoutingModule } from './academics-routing.module';
-import { CoursesHomeComponent } from './courses-home/courses-home.component';
+import { CoursesHomeComponent } from './course-catalog/course-catalog.component';
 import { MatIconModule } from '@angular/material/icon';
 import { OfferingsComponent } from './offerings/offerings.component';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatSelectModule } from '@angular/material/select';
 import { ReactiveFormsModule } from '@angular/forms';
+import { AcademicsHomeComponent } from './academics-home/academics-home.component';
 
 @NgModule({
-  declarations: [CoursesHomeComponent, OfferingsComponent],
+  declarations: [
+    CoursesHomeComponent,
+    OfferingsComponent,
+    AcademicsHomeComponent
+  ],
   imports: [
     AcademicsRoutingModule,
     CommonModule,
