+++ frontend/src/app/academics/course-catalog/course-catalog.component.css
